Renamed Intermediate class & removed the commented out code
