+++ src/main/java/org/owasp/benchmark/testcode/BenchmarkTest02741.java
@@ -17,7 +17,7 @@
  */
 package org.owasp.benchmark.testcode;
 
-import java.io.*;
+
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -25,15 +25,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 @WebServlet(value = "/xss-00/BenchmarkTest02741")
-public class BenchmarkTest02741 extends Intermediate {
+public class BenchmarkTest02741 extends SanitizingHttpServlet {
 
     private static final long serialVersionUID = 1L;
-
-    /*@Override
-    public void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        doPost(request, response);
-    }*/
 
     @Override
     public void doPost(HttpServletRequest request, HttpServletResponse response)
@@ -57,12 +51,4 @@ public class BenchmarkTest02741 extends Intermediate {
 
         response.getWriter().println(htmlResponse);
     }
-
-    /*protected String goodSanitationForIdAttribute(String parameter) {
-        String parameter1 = parameter;
-        parameter1 = parameter1.replaceAll("'", "&#39;");
-        parameter1 = parameter1.replaceAll("\"", "&quot;");
-        parameter1 = parameter1.replaceAll("&", "&amp;");
-        return parameter1;
-    }*/
 }
