+++ src/main/java/org/owasp/benchmark/testcode/BenchmarkTest02745.java
@@ -24,15 +24,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 @WebServlet(value = "/xss-00/BenchmarkTest02745")
-public class BenchmarkTest02745 extends Intermediate {
+public class BenchmarkTest02745 extends SanitizingHttpServlet {
 
     private static final long serialVersionUID = 1L;
-
-    /*@Override
-    public void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        doPost(request, response);
-    }*/
 
     @Override
     public void doPost(HttpServletRequest request, HttpServletResponse response)
@@ -50,12 +44,4 @@ public class BenchmarkTest02745 extends Intermediate {
 
         response.getWriter().println(htmlResponse);
     }
-
-    /*protected String goodSanitationForIdAttribute(String parameter) {
-        String parameter1 = parameter;
-        parameter1 = parameter1.replaceAll("'", "&#39;");
-        parameter1 = parameter1.replaceAll("\"", "&quot;");
-        parameter1 = parameter1.replaceAll("&", "&amp;");
-        return parameter1;
-    }*/
 }
